Update __init__.py
The audio input stream of one of the stations in my DAB ensemble is frequently causing a 'FATAL state, too many start retries too quickly' entry in the supervisord log. I suggest to  overcome this by adding a higher number of startretries to the default supervisor parameters for the audioencoder.